merge relative paths in VFS on windows without erroring
splitpath on windows results in a defined null-length string for the
volume of portion of the path every time. When merging relative paths onto
the root path of a VFS, Badger broke because it only checked the volume
part for definedness.

This fix allows Badger (and thus its dependencies) to install on Windows
again.